A whole bunch of duplicated code with ambiguous casts from int to float or double; the code duplication is somewhat worrying. Shouldn't this be one single macro or template in a common header?
svn path=/trunk/KDE/kdeaccessibility/kttsd/; revision=836767

+++ kcmkttsmgr/kcmkttsmgr.cpp
@@ -2002,13 +2002,13 @@ void KCMKttsMgr::keepAudioCheckBox_toggled(bool checked)
 // with alpha = 1000/(log(200)-log(50))
 
 int KCMKttsMgr::percentToSlider(int percentValue) {
-    double alpha = 1000 / (log(200) - log(50));
-    return (int)floor (0.5 + alpha * (log(percentValue)-log(50)));
+    double alpha = 1000 / (log(200.0) - log(50.0));
+    return (int)floor (0.5 + alpha * (log((float)percentValue)-log(50.0)));
 }
 
 int KCMKttsMgr::sliderToPercent(int sliderValue) {
-    double alpha = 1000 / (log(200) - log(50));
-    return (int)floor(0.5 + exp (sliderValue/alpha + log(50)));
+    double alpha = 1000 / (log(200.0) - log(50.0));
+    return (int)floor(0.5 + exp (sliderValue/alpha + log(50.0)));
 }
 
 void KCMKttsMgr::timeBox_valueChanged(int percentValue) {
