+++ plugins/hadifix/hadifixconf.cpp
@@ -77,13 +77,13 @@ HadifixConfPrivate::~HadifixConfPrivate() {
 // with alpha = 1000/(log(200)-log(50))
 
 int HadifixConfPrivate::percentToSlider (int percentValue) {
-    double alpha = 1000 / (log(200) - log(50));
-    return (int)floor (0.5 + alpha * (log(percentValue)-log(50)));
+    double alpha = 1000 / (log(200.0) - log(50.0));
+    return (int)floor (0.5 + alpha * (log((float)percentValue)-log(50.0)));
 }
 
 int HadifixConfPrivate::sliderToPercent (int sliderValue) {
-    double alpha = 1000 / (log(200) - log(50));
-    return (int)floor(0.5 + exp (sliderValue/alpha + log(50)));
+    double alpha = 1000 / (log(200.0) - log(50.0));
+    return (int)floor(0.5 + exp (sliderValue/alpha + log(50.0)));
 }
 
 void HadifixConfPrivate::volumeBox_valueChanged (int percentValue) {
